Develop

+++ app/src/main/java/com/example/tapago/domain/model/Sheet.kt
@@ -2,7 +2,7 @@ package com.example.tapago.domain.model
 
 data class Sheet (
 
-    val idSheet: Long,
+    val idSheet: Int,
     val nameSheet: String,
     val qtdExercise: Int
 )
